refactor
* reorganize package
* properly fix clippy warning for when tracing is disabled.
